pyproject.toml
[project]
name = "nexus7-enterprise-ai"
version = "1.0.0"
description = "Beyond the Prompt: Enterprise AI Demo - Blade Runner Edition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.32.0",
    "snowflake-connector-python>=3.6.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "cryptography>=41.0.0",
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
